+++ src/main/java/com/favoriteplace/app/converter/CommentConverter.java
@@ -3,7 +3,7 @@ package com.favoriteplace.app.converter;
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.domain.community.Comment;
 import com.favoriteplace.app.domain.community.GuestBook;
-import com.favoriteplace.app.dto.UserInfoResponseDto;
+import com.favoriteplace.app.member.controller.dto.UserInfoResponseDto;
 import com.favoriteplace.app.dto.community.comment.CommentParentResponseDto;
 import com.favoriteplace.app.dto.community.guestbook.GuestBookResponseDto;
 import com.favoriteplace.app.dto.community.PostResponseDto;
