+++ src/main/java/com/favoriteplace/app/dto/community/comment/CommentParentResponseDto.java
@@ -1,6 +1,6 @@
 package com.favoriteplace.app.dto.community.comment;
 
-import com.favoriteplace.app.dto.UserInfoResponseDto;
+import com.favoriteplace.app.member.controller.dto.UserInfoResponseDto;
 import java.util.List;
 
 public record CommentParentResponseDto(UserInfoResponseDto userInfo, Long id, String content, String passedTime,
