+++ src/main/java/com/favoriteplace/app/item/domain/PointHistory.java
@@ -1,13 +1,19 @@
-package com.favoriteplace.app.domain.item;
+package com.favoriteplace.app.item.domain;
+
+import com.favoriteplace.app.member.domain.Member;
+import com.favoriteplace.app.member.domain.enums.PointType;
 
 import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import java.time.LocalDateTime;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -17,26 +23,28 @@ import lombok.AllArgsConstructor;
 import static lombok.AccessLevel.PRIVATE;
 import static lombok.AccessLevel.PROTECTED;
 
-import com.favoriteplace.app.member.domain.Member;
-import com.favoriteplace.app.domain.common.BaseTimeEntity;
-
 @Entity
 @Getter
 @Builder
 @NoArgsConstructor(access = PROTECTED)
 @AllArgsConstructor(access = PRIVATE)
-public class AcquiredItem extends BaseTimeEntity {
+public class PointHistory {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "acquired_item_id")
+    @Column(name = "point_history_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
     private Member member;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "item_id", nullable = false)
-    private Item item;
+    @Column(nullable = false)
+    private long point;
 
+    @Column(nullable = false)
+    private LocalDateTime dealtAt;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private PointType pointType;
 }
