+++ src/main/java/com/favoriteplace/app/item/repository/PointHistoryRepository.java
@@ -1,6 +1,6 @@
-package com.favoriteplace.app.repository;
+package com.favoriteplace.app.item.repository;
 
-import com.favoriteplace.app.domain.item.PointHistory;
+import com.favoriteplace.app.item.domain.PointHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
