+++ src/main/java/com/favoriteplace/app/member/controller/MemberController.java
@@ -1,12 +1,12 @@
 package com.favoriteplace.app.member.controller;
 
-import com.favoriteplace.app.dto.member.KaKaoSignUpRequestDto;
-import com.favoriteplace.app.dto.member.MemberDto;
-import com.favoriteplace.app.dto.member.MemberDto.EmailCheckReqDto;
-import com.favoriteplace.app.dto.member.MemberDto.EmailDuplicateResDto;
-import com.favoriteplace.app.dto.member.MemberDto.EmailSendResDto;
-import com.favoriteplace.app.dto.member.MemberDto.MemberSignUpReqDto;
-import com.favoriteplace.app.dto.member.MemberDto.TokenInfo;
+import com.favoriteplace.app.member.controller.dto.KaKaoSignUpRequestDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto.EmailCheckReqDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto.EmailDuplicateResDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto.EmailSendResDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto.MemberSignUpReqDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto.TokenInfo;
 import com.favoriteplace.app.member.service.MailSendService;
 import com.favoriteplace.app.member.service.MemberService;
 import com.favoriteplace.global.security.provider.JwtTokenProvider;
