+++ src/main/java/com/favoriteplace/app/member/controller/dto/AuthKakaoLoginDto.java
@@ -1,4 +1,4 @@
-package com.favoriteplace.app.dto.member;
+package com.favoriteplace.app.member.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
