+++ src/main/java/com/favoriteplace/app/member/controller/dto/KaKaoSignUpRequestDto.java
@@ -1,9 +1,9 @@
-package com.favoriteplace.app.dto.member;
+package com.favoriteplace.app.member.controller.dto;
 
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.member.domain.enums.LoginType;
 import com.favoriteplace.app.member.domain.enums.MemberStatus;
-import com.favoriteplace.app.domain.item.Item;
+import com.favoriteplace.app.item.domain.Item;
 
 public record KaKaoSignUpRequestDto(
         String nickname,
