+++ src/main/java/com/favoriteplace/app/member/controller/dto/MemberDto.java
@@ -1,9 +1,9 @@
-package com.favoriteplace.app.dto.member;
+package com.favoriteplace.app.member.controller.dto;
 
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.member.domain.enums.LoginType;
 import com.favoriteplace.app.member.domain.enums.MemberStatus;
-import com.favoriteplace.app.domain.item.Item;
+import com.favoriteplace.app.item.domain.Item;
 import jakarta.validation.constraints.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
