+++ src/main/java/com/favoriteplace/app/member/domain/Member.java
@@ -3,7 +3,7 @@ package com.favoriteplace.app.member.domain;
 import com.favoriteplace.app.domain.common.BaseTimeEntity;
 import com.favoriteplace.app.member.domain.enums.LoginType;
 import com.favoriteplace.app.member.domain.enums.MemberStatus;
-import com.favoriteplace.app.domain.item.Item;
+import com.favoriteplace.app.item.domain.Item;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
