+++ src/main/java/com/favoriteplace/app/member/service/MemberService.java
@@ -6,15 +6,15 @@ import static com.favoriteplace.global.exception.ErrorCode.USER_ALREADY_EXISTS;
 import static com.favoriteplace.global.exception.ErrorCode.USER_NOT_FOUND;
 
 import com.favoriteplace.app.member.domain.Member;
-import com.favoriteplace.app.domain.item.Item;
-import com.favoriteplace.app.dto.UserInfoResponseDto;
-import com.favoriteplace.app.dto.member.AuthKakaoLoginDto;
-import com.favoriteplace.app.dto.member.KaKaoSignUpRequestDto;
-import com.favoriteplace.app.dto.member.MemberDto;
-import com.favoriteplace.app.dto.member.MemberDto.EmailDuplicateResDto;
-import com.favoriteplace.app.dto.member.MemberDto.EmailSendReqDto;
-import com.favoriteplace.app.dto.member.MemberDto.MemberSignUpReqDto;
-import com.favoriteplace.app.repository.ItemRepository;
+import com.favoriteplace.app.item.domain.Item;
+import com.favoriteplace.app.member.controller.dto.UserInfoResponseDto;
+import com.favoriteplace.app.member.controller.dto.AuthKakaoLoginDto;
+import com.favoriteplace.app.member.controller.dto.KaKaoSignUpRequestDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto.EmailDuplicateResDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto.EmailSendReqDto;
+import com.favoriteplace.app.member.controller.dto.MemberDto.MemberSignUpReqDto;
+import com.favoriteplace.app.item.repository.ItemRepository;
 import com.favoriteplace.app.member.repository.MemberRepository;
 import com.favoriteplace.global.exception.RestApiException;
 import com.favoriteplace.global.s3Image.AmazonS3ImageManager;
