+++ src/main/java/com/favoriteplace/app/service/MyPageCommandService.java
@@ -4,13 +4,13 @@ import com.favoriteplace.app.converter.CommonConverter;
 import com.favoriteplace.app.domain.Block;
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.domain.enums.ItemType;
-import com.favoriteplace.app.domain.item.Item;
+import com.favoriteplace.app.item.domain.Item;
 import com.favoriteplace.app.dto.CommonResponseDto;
 import com.favoriteplace.app.dto.MyPageDto;
 import com.favoriteplace.app.dto.MyPageDto.MyFcmTokenDto;
-import com.favoriteplace.app.repository.AcquiredItemRepository;
+import com.favoriteplace.app.item.repository.AcquiredItemRepository;
 import com.favoriteplace.app.repository.BlockRepository;
-import com.favoriteplace.app.repository.ItemRepository;
+import com.favoriteplace.app.item.repository.ItemRepository;
 import com.favoriteplace.app.member.repository.MemberRepository;
 import com.favoriteplace.app.notification.service.FCMNotificationService;
 import com.favoriteplace.global.exception.ErrorCode;
