+++ src/main/java/com/favoriteplace/app/service/MyPageQueryService.java
@@ -5,12 +5,12 @@ import com.favoriteplace.app.domain.Block;
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.domain.enums.ItemType;
 import com.favoriteplace.app.domain.enums.SaleStatus;
-import com.favoriteplace.app.domain.item.AcquiredItem;
+import com.favoriteplace.app.item.domain.AcquiredItem;
 import com.favoriteplace.app.domain.travel.CompleteRally;
 import com.favoriteplace.app.domain.travel.LikedRally;
 import com.favoriteplace.app.domain.travel.Rally;
 import com.favoriteplace.app.dto.MyPageDto;
-import com.favoriteplace.app.repository.AcquiredItemRepository;
+import com.favoriteplace.app.item.repository.AcquiredItemRepository;
 import com.favoriteplace.app.repository.BlockRepository;
 import com.favoriteplace.app.repository.CommentRepository;
 import com.favoriteplace.app.repository.CompleteRallyRepository;
