+++ src/main/java/com/favoriteplace/app/service/PilgrimageCommandService.java
@@ -1,12 +1,12 @@
 package com.favoriteplace.app.service;
 
 import com.favoriteplace.app.converter.CommonConverter;
-import com.favoriteplace.app.converter.PointHistoryConverter;
+import com.favoriteplace.app.item.converter.PointHistoryConverter;
 import com.favoriteplace.app.member.domain.Member;
 import com.favoriteplace.app.domain.community.GuestBook;
 import com.favoriteplace.app.member.domain.enums.PointType;
 import com.favoriteplace.app.domain.enums.RallyVersion;
-import com.favoriteplace.app.domain.item.AcquiredItem;
+import com.favoriteplace.app.item.domain.AcquiredItem;
 import com.favoriteplace.app.domain.travel.CompleteRally;
 import com.favoriteplace.app.domain.travel.LikedRally;
 import com.favoriteplace.app.domain.travel.Pilgrimage;
@@ -15,12 +15,12 @@ import com.favoriteplace.app.domain.travel.VisitedPilgrimage;
 import com.favoriteplace.app.dto.CommonResponseDto;
 import com.favoriteplace.app.dto.travel.PilgrimageCertifyRequestDto;
 import com.favoriteplace.app.dto.travel.PilgrimageSocketDto;
-import com.favoriteplace.app.repository.AcquiredItemRepository;
+import com.favoriteplace.app.item.repository.AcquiredItemRepository;
 import com.favoriteplace.app.repository.CompleteRallyRepository;
 import com.favoriteplace.app.repository.GuestBookRepository;
 import com.favoriteplace.app.repository.LikedRallyRepository;
 import com.favoriteplace.app.repository.PilgrimageRepository;
-import com.favoriteplace.app.repository.PointHistoryRepository;
+import com.favoriteplace.app.item.repository.PointHistoryRepository;
 import com.favoriteplace.app.repository.RallyRepository;
 import com.favoriteplace.app.repository.VisitedPilgrimageRepository;
 import com.favoriteplace.app.notification.service.FCMNotificationService;
